invert int and float in args

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -6,7 +6,7 @@ integers and floats and returns their sum as a float.
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float | int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int | float]]) -> float:
     """
     Function to sum a list of integers and floats
     Args:
